requirements-windows.txt
mss
Pillow
winocr
pyinstaller
